Refactor CDN client
