+++ SteamKit2/Tests/CDNClientFacts.cs
@@ -4,6 +4,7 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using SteamKit2;
+using SteamKit2.CDN;
 using Xunit;
 
 namespace Tests
@@ -11,27 +12,68 @@ namespace Tests
     public class CDNClientFacts
     {
         [Fact]
-        public async Task ThrowsSteamKitWebExceptionOnUnsuccessfulWebResponse()
+        public async Task ThrowsSteamKitWebExceptionOnUnsuccessfulWebResponseForManifest()
         {
-            var configuration = SteamConfiguration.Create(x => x.WithHttpClientFactory(() => new HttpClient(new TeapotHttpMessageHandler())));
-            var steam = new SteamClient(configuration);
-            var client = new CDNClient(steam);
+            var configuration = SteamConfiguration.Create( x => x.WithHttpClientFactory( () => new HttpClient( new TeapotHttpMessageHandler() ) ) );
+            var steam = new SteamClient( configuration );
+            var client = new Client( steam );
+            var server = new Server
+            {
+                Protocol = Server.ConnectionProtocol.HTTP,
+                Host = "localhost",
+                VHost = "localhost",
+                Port = 80
+            };
 
-            try
+            var ex = await Assert.ThrowsAsync<SteamKitWebRequestException>( () => client.DownloadManifestAsync( depotId: 0, manifestId: 0, manifestRequestCode: 0, server ) );
+            Assert.Equal( ( HttpStatusCode )418, ex.StatusCode );
+        }
+
+        [Fact]
+        public async Task ThrowsSteamKitWebExceptionOnUnsuccessfulWebResponseForChunk()
+        {
+            var configuration = SteamConfiguration.Create( x => x.WithHttpClientFactory( () => new HttpClient( new TeapotHttpMessageHandler() ) ) );
+            var steam = new SteamClient( configuration );
+            var client = new Client( steam );
+            var server = new Server
             {
-                await client.DownloadManifestAsync(0, 0, "localhost", "12345");
-                throw new InvalidOperationException("This should be unreachable.");
-            }
-            catch (SteamKitWebRequestException ex)
+                Protocol = Server.ConnectionProtocol.HTTP,
+                Host = "localhost",
+                VHost = "localhost",
+                Port = 80
+            };
+            var chunk = new DepotManifest.ChunkData
             {
-                Assert.Equal((HttpStatusCode)418, ex.StatusCode);
-            }
+                ChunkID = new byte[] { 0xFF },
+            };
+
+            var ex = await Assert.ThrowsAsync<SteamKitWebRequestException>( () => client.DownloadDepotChunkAsync( depotId: 0, chunk, server ) );
+            Assert.Equal( ( HttpStatusCode )418, ex.StatusCode );
+        }
+
+        [Fact]
+        public async Task ThrowsWhenNoChunkIDIsSet()
+        {
+            var configuration = SteamConfiguration.Create( x => x.WithHttpClientFactory( () => new HttpClient( new TeapotHttpMessageHandler() ) ) );
+            var steam = new SteamClient( configuration );
+            var client = new Client( steam );
+            var server = new Server
+            {
+                Protocol = Server.ConnectionProtocol.HTTP,
+                Host = "localhost",
+                VHost = "localhost",
+                Port = 80
+            };
+            var chunk = new DepotManifest.ChunkData();
+
+            var ex = await Assert.ThrowsAsync<ArgumentException>( () => client.DownloadDepotChunkAsync( depotId: 0, chunk, server ) );
+            Assert.Equal( "chunk", ex.ParamName );
         }
 
         sealed class TeapotHttpMessageHandler : HttpMessageHandler
         {
-            protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-                => Task.FromResult(new HttpResponseMessage((HttpStatusCode)418));
+            protected override Task<HttpResponseMessage> SendAsync( HttpRequestMessage request, CancellationToken cancellationToken )
+                => Task.FromResult( new HttpResponseMessage( ( HttpStatusCode )418 ) );
         }
     }
 }
